refactor comment module
