--- src/AddComment/style.css
@@ -1,6 +1,0 @@
-.AddComment-message {
-  padding: 15px;
-  border-radius: 3px;
-  margin-bottom: 10px;
-  font-size: 12px;
-}
