+++ src/Comment/CommentAdd/index.js
@@ -3,13 +3,11 @@ import { Mutation } from 'react-apollo';
 
 import { ADD_COMMENT } from './mutations';
 
-import TextArea from '../TextArea';
-import Button from '../Button';
-import ErrorMessage from '../Error';
+import TextArea from '../../TextArea';
+import Button from '../../Button';
+import ErrorMessage from '../../Error';
 
-import './style.css';
-
-class AddComment extends Component {
+class CommentAdd extends Component {
   state = {
     value: '',
   };
@@ -52,4 +50,4 @@ class AddComment extends Component {
   }
 }
 
-export default AddComment;
+export default CommentAdd;
