+++ src/Comment/CommentItem/style.css
@@ -1,0 +1,4 @@
+.Comment {
+  display: flex;
+  align-items: baseline;
+}
