+++ src/Comment/index.js
@@ -1,0 +1,3 @@
+import CommentList from './CommentList';
+
+export default CommentList;
